Split layer into submodules.
